refactor: extracts Subject from Subjects
